fix(tests): keep standalone scripts import-safe
I now guard argv-driven and environment-dependent script entry points during import, preserve their executable behavior, and keep the complete pytest suite free of collection side effects and timing assumptions.
