+++ scripts/tests/test_qualified_identity_literal_rule.py
@@ -12,11 +12,7 @@ from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[2]
 RULE = ROOT / "rules/rust/authority/no-qualified-identity-literal-comparison.yml"
-AST_GREP = (
-    Path(sys.argv[1]).resolve()
-    if len(sys.argv) > 1
-    else ROOT / ".ast-grep/tool/bin/ast-grep"
-)
+AST_GREP = ROOT / ".ast-grep/tool/bin/ast-grep"
 
 
 def findings(source: str) -> list[dict[str, object]]:
@@ -46,20 +42,29 @@ def findings(source: str) -> list[dict[str, object]]:
         return [json.loads(line) for line in result.stdout.splitlines() if line]
 
 
-red = findings(
-    'fn is_rc_new(name: &str) -> bool { name == "Rc::new" }\n'
-    'fn is_monitor_id(name: &str) -> bool { "MonitorRef::id" != name }\n'
-)
-if len(red) != 2:
-    raise SystemExit(
-        f"qualified identity literal counterfactual found {len(red)}, want 2: {red}"
+def main() -> None:
+    global AST_GREP
+    if len(sys.argv) > 1:
+        AST_GREP = Path(sys.argv[1]).resolve()
+
+    red = findings(
+        'fn is_rc_new(name: &str) -> bool { name == "Rc::new" }\n'
+        'fn is_monitor_id(name: &str) -> bool { "MonitorRef::id" != name }\n'
     )
+    if len(red) != 2:
+        raise SystemExit(
+            f"qualified identity literal counterfactual found {len(red)}, want 2: {red}"
+        )
 
-green = findings(
-    "fn is_rc_new(family: RuntimeCallFamily) -> bool { "
-    "family == RuntimeCallFamily::RcNew }\n"
-)
-if green:
-    raise SystemExit(f"typed identity counterfactual was flagged: {green}")
+    green = findings(
+        "fn is_rc_new(family: RuntimeCallFamily) -> bool { "
+        "family == RuntimeCallFamily::RcNew }\n"
+    )
+    if green:
+        raise SystemExit(f"typed identity counterfactual was flagged: {green}")
 
-print("qualified identity literal comparison counterfactuals: PASS")
+    print("qualified identity literal comparison counterfactuals: PASS")
+
+
+if __name__ == "__main__":
+    main()
